Added frontend

+++ bankingapp/src/main/java/pl/Portfolio/bankingapp/Services/AccountService.java
@@ -2,9 +2,12 @@ package pl.Portfolio.bankingapp.Services;
 
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 import pl.Portfolio.bankingapp.DTO.AccountDto;
+import pl.Portfolio.bankingapp.Exceptions.AccountNotFoundException;
 import pl.Portfolio.bankingapp.Exceptions.NegativeBalanceException;
+import pl.Portfolio.bankingapp.Exceptions.UserNotFoundException;
 import pl.Portfolio.bankingapp.Model.Account;
 import pl.Portfolio.bankingapp.Model.User;
 import pl.Portfolio.bankingapp.Repository.AccountRepository;
@@ -20,9 +23,11 @@ public class AccountService {
     private final UserRepository userRepository;
     private ModelMapper modelMapper = new ModelMapper();
 
-    private String getUserNameById(int id) {
-        User user = userRepository.getById(id);
-        return user.getUsername();
+    private String getUserNameById(int id) throws UserNotFoundException {
+        if (userRepository.getById(id) == null)
+            throw new UserNotFoundException("User with id " + id + " not found");
+
+        return userRepository.getById(id).getUsername();
     }
 
     private AccountDto mapAccountToAccountDto(Account account) {
@@ -43,12 +48,10 @@ public class AccountService {
         this.userRepository = userRepository;
     }
 
-    public int saveAccounts(List<AccountDto> accountsDto) {
-        List<Account> accounts = new ArrayList<>();
-        for (AccountDto accountDto : accountsDto) {
-            accounts.add(mapAccountDtoToAccount(accountDto));
-        }
-        return accountRepository.save(accounts);
+    public int saveAccount(AccountDto accountDto) {
+        Account account = mapAccountDtoToAccount(accountDto);
+
+        return accountRepository.save(account);
     }
 
     public List<AccountDto> getAllAccounts() {
@@ -60,63 +63,74 @@ public class AccountService {
         return accountsDto;
     }
 
-    public AccountDto getAccountById(int id) {
+    public AccountDto getAccountById(int id) throws AccountNotFoundException {
         Account account = accountRepository.getById(id);
+
+        if (account == null)
+            throw new AccountNotFoundException("Account with id " + id + " not found");
+
         return mapAccountToAccountDto(account);
     }
 
-    public int updateAccount(int id, AccountDto accountDto) throws NegativeBalanceException{
+    public int updateAccount(int id, AccountDto accountDto) throws NegativeBalanceException, AccountNotFoundException{
         Account account = accountRepository.getById(id);
+
+        if (account == null)
+            throw new AccountNotFoundException("Account with id " + id + " not found");
 
         if (accountDto.getBalance() < 0) {
             throw new NegativeBalanceException("Balance cannot be negative");
         }
 
-        if (account != null) {
-            account.setAccount_number(accountDto.getAccount_number());
-            account.setAccount_type(accountDto.getAccount_type());
-            account.setBalance(accountDto.getBalance());
+        account.setAccount_number(accountDto.getAccount_number());
+        account.setAccount_type(accountDto.getAccount_type());
+        account.setBalance(accountDto.getBalance());
 
-            accountRepository.update(account);
+        accountRepository.update(account);
 
-            return 1;
-        } else {
-            return -1;
-        }
+        return 1;
     }
 
-    public int partiallyUpdateAccount(int id, AccountDto accountDto) throws NegativeBalanceException {
+    public int partiallyUpdateAccount(int id, AccountDto accountDto) throws NegativeBalanceException, AccountNotFoundException {
         Account account = accountRepository.getById(id);
+
+        if (account == null)
+            throw new AccountNotFoundException("Account with id " + id + " not found");
 
         if (accountDto.getBalance() < 0) {
             throw new NegativeBalanceException("Balance cannot be negative");
         }
 
-        if (account != null) {
-            if (accountDto.getAccount_number() != null) {
-                account.setAccount_number(accountDto.getAccount_number());
-            }
-            if (accountDto.getAccount_type() != null) {
-                account.setAccount_type(accountDto.getAccount_type());
-            }
-            if (accountDto.getBalance() >= 0) {
-                account.setBalance(accountDto.getBalance());
-            }
-
-            accountRepository.update(account);
-
-            return 1;
-        } else {
-            return -1;
+        if (accountDto.getAccount_number() != null) {
+            account.setAccount_number(accountDto.getAccount_number());
         }
+        if (accountDto.getAccount_type() != null) {
+            account.setAccount_type(accountDto.getAccount_type());
+        }
+        if (accountDto.getBalance() >= 0) {
+            account.setBalance(accountDto.getBalance());
+        }
+
+        accountRepository.update(account);
+
+        return 1;
     }
 
     public int deleteAccount(int id) {
+        if (accountRepository.getById(id) == null)
+            throw new AccountNotFoundException("Account with id " + id + " not found");
+
         return accountRepository.delete(id);
     }
 
-    public List<AccountDto> getAccountsByOwnerId(int ownerId) {
+    public List<AccountDto> getAccountsByOwnerId(int ownerId) throws UserNotFoundException, AccountNotFoundException{
         List<Account> accounts = accountRepository.findByUserId(ownerId);
+
+        if (userRepository.getById(ownerId) == null)
+            throw new UserNotFoundException("User with id " + ownerId + " not found");
+
+        if (accounts == null)
+            throw new AccountNotFoundException("Account with owner id " + ownerId + " not found");
 
         List<AccountDto> accountsDto = new ArrayList<>();
         for (Account account : accounts) {
